requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic[email]>=2.0.0  # For email validation
python-dotenv>=1.0.0
openai>=1.0.0
python-multipart>=0.0.5  # For handling file uploads
google-cloud-vision>=3.4.0  # For OCR using Google Vision API
Pillow>=9.0.0  # For image processing
motor>=3.1.1  # MongoDB async driver
pymongo>=4.3.3  # MongoDB driver
passlib>=1.7.4  # For password hashing
python-jose>=3.3.0  # For JWT tokens
bcrypt>=4.0.1  # For password hashing
